feat(connect): add echo example
